+++ .eslintrc.js
@@ -2,6 +2,7 @@ module.exports = {
   env: {
     browser: true,
     es2021: true,
+    node: true,
   },
 
   extends: ["eslint:recommended", "plugin:react/recommended", "prettier"],
