+++ src/pages/orderList/ReturnView.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const ReturnViewCom = () => {
+const ReturnView = () => {
   return (
     <div>
       <p>환품.반품페이지 테스트</p>
@@ -8,4 +8,4 @@ const ReturnViewCom = () => {
   );
 };
 
-export default ReturnViewCom;
+export default ReturnView;
